+++ src/app/exportAtsumaru.ts
@@ -32,6 +32,13 @@ export function promiseExportAtsumaru(param: ExportHTMLParameterObject): Promise
 			if (!gameJson.environment) {
 				gameJson.environment = {};
 			}
+			if (!gameJson.environment.niconico || !gameJson.environment.niconico.supportedModes) {
+				// モード指定がなければ、常に指定可能なモードであるsingleモードを追加する。
+				completedParam.logger.warn("Supported-modes must be specified, so \"single\" has been added");
+				gameJson.environment.niconico = {
+					"supportedModes": ["single"]
+				};
+			}
 			if (!gameJson.environment.external) {
 				gameJson.environment.external = {};
 			}